+++ backend/gestion/src/main/java/com/entreprise/gestion/rh/service/DiplomeFiliereService.java
@@ -15,6 +15,6 @@ public class DiplomeFiliereService {
     private BesoinDiplomeFiliereRepository besoinDiplomeFiliereRepository;
 
     public List<BesoinDiplomeFiliere> findAll() { return besoinDiplomeFiliereRepository.findAll(); }
-    public Optional<BesoinDiplomeFiliere> findById(Long id) { return besoinDiplomeFiliereRepository.findById(id); }
+    public Optional<BesoinDiplomeFiliere> findById(Integer id) { return besoinDiplomeFiliereRepository.findById(id); }
     public BesoinDiplomeFiliere save(BesoinDiplomeFiliere bdf) { return besoinDiplomeFiliereRepository.save(bdf); }
 }
